Load events in batches of 1000 when processing stream

+++ framework-tools-replay/src/main/java/uk/gov/justice/framework/tools/replay/ProgressChecker.java
@@ -4,7 +4,7 @@ public class ProgressChecker {
 
     private static final int PROGRESS_INTERVAL = 100;
 
-    public boolean shouldLogProgress(final int index) {
+    public boolean shouldLogProgress(final long index) {
         return index % PROGRESS_INTERVAL == 0;
     }
 }
